[Refactor] 이메일 로그인 컴포즈 적용 시작 1. 이메일 로그인 화면의 '로그인 버튼' 과 이름이 겹치지 않도록 소셜로그인의 이메일로 가기 버튼명 변경(EmailLoginButton -> NavigateToEmailLoginButton) 2. 이메일 로그인 스크린 생성 3. onCreateView에서 컴포즈뷰로 생성한 스크린을 리턴 4. 컴포넌트 패키지 생성
