+++ app/src/main/java/kr/co/lion/modigm/ui/login/email/EmailLoginScreen.kt
@@ -1,0 +1,16 @@
+package kr.co.lion.modigm.ui.login.email
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun EmailLoginScreen(
+    viewModel: EmailLoginViewModel
+) {
+    Box(
+        modifier = Modifier
+            .fillMaxSize(),
+    )
+}
